Task-240: Start model refactor to use common classes
