+++ src/main/java/taiga/model/query/common/OwnerExtraInfo.java
@@ -1,5 +1,5 @@
 
-package taiga.model.query.userstories;
+package taiga.model.query.common;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
