+++ src/main/java/taiga/model/query/common/ProjectExtraInfo.java
@@ -1,14 +1,18 @@
 
-package taiga.model.query.userstories;
+package taiga.model.query.common;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class Project {
+public class ProjectExtraInfo {
 
     @SerializedName("id")
     @Expose
     private Integer id;
+
+    @SerializedName("logo_small_url")
+    @Expose
+    private Object logoSmallUrl;
 
     @SerializedName("name")
     @Expose
@@ -24,6 +28,14 @@ public class Project {
 
     public void setId(Integer id) {
         this.id = id;
+    }
+
+    public Object getLogoSmallUrl() {
+        return logoSmallUrl;
+    }
+
+    public void setLogoSmallUrl(Object logoSmallUrl) {
+        this.logoSmallUrl = logoSmallUrl;
     }
 
     public String getName() {
