--- src/main/java/taiga/model/query/sprint/Project.java
@@ -1,43 +1,0 @@
-package taiga.model.query.sprint;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Project {
-
-    @SerializedName("id")
-    @Expose
-    private Integer id;
-
-    @SerializedName("name")
-    @Expose
-    private String name;
-
-    @SerializedName("slug")
-    @Expose
-    private String slug;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSlug() {
-        return slug;
-    }
-
-    public void setSlug(String slug) {
-        this.slug = slug;
-    }
-}
