+++ src/main/java/taiga/model/query/sprint/Sprint.java
@@ -5,6 +5,7 @@ import com.google.gson.annotations.SerializedName;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
+import taiga.model.query.common.ProjectExtraInfo;
 
 public class Sprint {
