+++ src/main/java/taiga/model/query/tasks/UserStoryExtraInfo.java
@@ -4,6 +4,7 @@ import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.List;
+import taiga.model.query.common.Epic;
 
 public class UserStoryExtraInfo {
