+++ src/main/java/taiga/model/query/userstories/UserStory.java
@@ -5,7 +5,11 @@ import java.util.Date;
 import java.util.List;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
-import taiga.model.query.sprint.Epic;
+import taiga.model.query.common.AssignedToExtraInfo;
+import taiga.model.query.common.Epic;
+import taiga.model.query.common.OwnerExtraInfo;
+import taiga.model.query.common.ProjectExtraInfo;
+import taiga.model.query.common.StatusExtraInfo;
 
 public class UserStory implements UserStoryInterface {
